chore: rename controller.js to custom-controls.js
